refactor(home): add a mapper
