+++ FitTrack_iOS/FitTrack_iOS/Presentation/Mappers/TrainingDomainToItemMapper.swift
@@ -1,0 +1,32 @@
+//
+//  TrainingDomainToItemMapper.swift
+//  FitTrack_iOS
+//
+//  Created by Ariana Rodríguez on 28/09/25.
+//
+
+import Foundation
+
+struct TrainingDomainToItemMapper {
+    func map(_ domain: Training, user: User) -> TrainingItem {
+        let isoFormatter = ISO8601DateFormatter()
+        isoFormatter.formatOptions = [.withInternetDateTime]
+        isoFormatter.timeZone = TimeZone(secondsFromGMT: .zero)
+        let date = isoFormatter.date(from: domain.scheduledAt ?? "") ?? Date()
+        let metric = ["2h", "1h", "45m"].randomElement()!
+        let icon = ["figure.soccer", "figure.strengthtraining.traditional", "figure.yoga"].randomElement()!
+        let userImage = ["sarah", "joey_t", "benito_bodoque", "person"].randomElement()!
+        
+        return .init(
+            id: domain.id ?? UUID(),
+            date: date,
+            metric: metric,
+            icon: icon,
+            userITem: UserItem(
+                id: UUID(uuidString: user.id ?? "") ?? UUID(),
+                image: userImage,
+                firstName: user.profile.name,
+                lastName: ""
+            ))
+    }
+}
