+++ FitTrack_iOS/FitTrack_iOS/Presentation/Views/Home /HomeView.swift
@@ -157,4 +157,3 @@ struct HomeView: View {
     HomeView()
         .environment(AppState())
 }
-
