+++ fittrack_server/Sources/fittrack/DTOs/UserDTO.swift
@@ -17,8 +17,7 @@ struct UserRegisterDTO: Content {
     
     
     enum CodingKeys: String, CodingKey {
-        case name, email, password, role
-        case coachID = "coach_id"
+        case email, password, role, profile
     }
     
     func toModel(withHashedPassword password: String) -> User {
